requirements.txt
bleak>=0.22.0
fastapi>=0.109.0
httpx>=0.27.0
pydantic>=2.5.0
python-chess>=1.10.0
uvicorn[standard]>=0.27.0
